requirements.txt
paho-mqtt==1.6.1
apscheduler==3.10.4
